Move DTextFormatter to TS
