+++ app/javascript/entrypoints/application.ts
@@ -7,7 +7,7 @@ Rails.start();
 
 // Common imports for all controllers
 import Autocomplete from "@/components/autocomplete";
-import DTextFormatter from "@/components/dtext_formatter";
+import DTextFormatter from "@/components/DTextFormatter";
 import "@/core/analytics";
 import "@/core/AuthOverlay";
 import Blacklist from "@/core/blacklists";
