+++ app/javascript/src/js/core/dtext_formatter_loader.js
@@ -1,4 +1,4 @@
-import DTextFormatter from "@/components/dtext_formatter";
+import DTextFormatter from "@/components/DTextFormatter";
 
 $(() => {
   for (const one of $(".dtext-formatter")) {
